Handled several exception that are permission denied missing entry from database

+++ validatortask.py
@@ -26,54 +26,66 @@ import sys
 import json
 import misc
 import migconfig
+import traceback
 
 def report_error(message):
 	global vclient
 	global mqclient
-	global procid
+	global id
 
 	vclient.close()
 	mqclient.close()
-	errormsg = procid + " Validator task report error: " + message
+	errormsg = id + " Validator task report error: " + message
 	misc.report_error(errormsg)
 
 def callback(ch, method, properties, body):
 	stats = "OK"
-	global procid
+	global id
 
+	procid = id + "-" + misc.generate_id(size=15)
 	file_logger.info(procid + " Start validation: " + repr(body))
 	dbmeta = logger.file_get(body)
 	file_logger.info(procid + " Database meta data: " + repr(dbmeta))
-	base = "/tmp"
-	name = os.path.basename(dbmeta["path"])
-	fullpath = os.path.join(base, name)
-	trans.download_file(key=dbmeta["path"], name=fullpath)
-	mogmeta = FileInfo("/tmp", name)
-	file_logger.info(procid + " MogileFS meta data: " + repr(mogmeta.to_collection()))
-	if mogmeta.equal_meta(dbmeta):
-		logger.file_validated(dbmeta)
+	
+	if dbmeta:
+		base = "/tmp"
+		name = os.path.basename(dbmeta["path"])
+		fullpath = os.path.join(base, name)
+		try:
+			if os.path.isfile(fullpath):
+				os.remove(fullpath)
+		except OSError:
+			suffix = misc.generate_id()
+			fullpath = fullpath + "-" + suffix
+		
+		trans.download_file(key=dbmeta["path"], name=fullpath)
+		mogmeta = FileInfo("/tmp", name)
+		file_logger.info(procid + " MogileFS meta data: " + repr(mogmeta.to_collection()))
+		if mogmeta.equal_meta(dbmeta):
+			logger.file_validated(dbmeta)
+		else:
+			logger.file_corrupted(dbmeta)
+			migration_job = {"path": dbmeta["path"], "base": dbmeta["base"]}
+			mqclient.send(migration_job)
+			msg = procid + " File corrupted. Sent new job to migration queue: " + json.dumps(dbmeta)
+			logger.warning(msg)
+			stats = "CORRUPTED"
+
+		os.remove(fullpath)
 	else:
-		logger.file_corrupted(dbmeta)
-		migration_job = {"path": dbmeta["path"], "base": dbmeta["base"]}
-		mqclient.send(migration_job)
-		msg = procid + " File corrupted. Sent new job to migration queue: " + json.dumps(dbmeta)
-		logger.warning(msg)
-		stats = "CORRUPTED"
-	os.remove(fullpath)
+		if not trans.key_exist(body):
+			file_logger.error("Missing key: " + body)
+		
 	ch.basic_ack(delivery_tag = method.delivery_tag)
 	file_logger.info("%s End validation %s: %s" %(procid, repr(body), stats))
 
 if __name__ == "__main__":
 	validation_task = True
-	procid = misc.generate_id()
+	id = misc.generate_id()
 	while validation_task:
 		try:
 			domain = migconfig.domain
 			trackers = migconfig.trackers
-			#if len(sys.argv) == 3:
-			#	domain = sys.argv[1] if sys.argv[1] else None
-			#	listoftracker = sys.argv[2] if sys.argv[2] else None
-			#	trackers = listoftracker.split(";")
 			trans = MogTransport(domain=domain, trackers=trackers)
 			vclient = RMQClient(queue=migconfig.validation_queue)
 			mqclient = RMQClient(queue=migconfig.migration_queue)
